requirements.txt
# Bad Debt Early-Warning API - Python 3.10+
# Install: pip install -r requirements.txt

fastapi>=0.115.0,<1.0
uvicorn>=0.30.0,<1.0
python-multipart>=0.0.9

pandas>=2.1.0,<3.0
numpy>=1.24.0,<3.0
scikit-learn==1.6.1
joblib>=1.3.0,<2.0
lightgbm>=4.0.0,<5.0
imbalanced-learn>=0.11.0,<1.0

SQLAlchemy>=2.0.0,<3.0
PyMySQL>=1.1.0,<2.0
python-dotenv>=1.0.0,<2.0
